feat: Name 테스트코드 작성
